+++ config/mocha.conf.js
@@ -1,4 +1,0 @@
-/* eslint-disable */
-global.expect = require('chai').expect;
-global.sinon = require('sinon-chai');
-